requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
fastapi==0.100.0
flake8==6.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pydantic==2.1.0
pydantic_core==2.4.0
pyflakes==3.0.1
PyJWT==2.8.0
pytest==7.4.0
python-dotenv==1.0.0
redis==4.6.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
